// openapi/vitrivr-engine/model/informationNeedDescription.ts
/**
 * vitrivr engine API
 * Rest API for the vitrivr engine project. Provides query (runtime) and extraction (ingestion) endpoints
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OperatorDescription } from './operatorDescription';
import { QueryContext } from './queryContext';
import { InputData } from './inputData';


export interface InformationNeedDescription { 
    inputs: { [key: string]: InputData; };
    operations: { [key: string]: OperatorDescription; };
    output: string;
    context: QueryContext;
}