// openapi/vitrivr-engine/model/inputData.ts
/**
 * vitrivr engine API
 * Rest API for the vitrivr engine project. Provides query (runtime) and extraction (ingestion) endpoints
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InputType } from './inputType';


export interface InputData { 
    type: InputType;
    comparison?: string;
}